Remove redundant provider classes
