+++ packages/rx_bloc_cli/lib/src/extensions/string_extensions.dart
@@ -1,0 +1,12 @@
+/// Utility methods for operations on strings
+extension StringExtensions on String {
+  /// Check if string matches a provided regex
+  bool matches({required String regex}) {
+    final regExp = RegExp(regex);
+    final match = regExp.matchAsPrefix(this);
+    return match != null && match.end == length;
+  }
+
+  /// Converts the string to boolean value
+  bool toBool() => toLowerCase() == true.toString();
+}
