+++ packages/rx_bloc_cli/lib/src/models/configurations/project_configuration.dart
@@ -1,3 +1,7 @@
+import 'package:rx_bloc_cli/src/extensions/string_extensions.dart';
+
+import '../errors/command_usage_exception.dart';
+
 /// Contains configuration data for the project
 class ProjectConfiguration {
   /// Constructor with projectName and organisation parameters
@@ -11,10 +15,41 @@ class ProjectConfiguration {
 
   /// Project name
   final String projectName;
+
   /// Organisation
   final String organisation;
+
   /// Organisation name
   final String organisationName;
+
   /// Organisation domain
   final String organisationDomain;
+}
+
+/// Validations for project configuration
+extension ProjectConfigurationValidations on ProjectConfiguration {
+  /// Validates the provided project name
+  static String validateProjectName(String name) {
+    final projectNameRegex = '[a-z_][a-z0-9_]*';
+    if (!name.matches(regex: projectNameRegex)) {
+      throw CommandUsageException(
+        '"$name" is not a valid package name.\n\n'
+        'See https://dart.dev/tools/pub/pubspec#name for more information.',
+      );
+    }
+    return name;
+  }
+
+  /// Validates the provided organisation
+  static String validateOrganisation(String orgName) {
+    final organisationRegex =
+        '^([A-Za-z]{2,6})(\\.(?!-)[A-Za-z0-9-]{1,63}(?<!-))+\$';
+    if (orgName.trim().isEmpty) {
+      throw CommandUsageException('No organisation name specified.');
+    }
+    if (!orgName.matches(regex: organisationRegex)) {
+      throw CommandUsageException('Invalid organisation name.');
+    }
+    return orgName;
+  }
 }
